docker

+++ root_tool/install_root_docker.sh
@@ -1,7 +1,14 @@
 #!/bin/bash
 
-function install_root_docker(){
-    echo "[...]install_root_docker......."
+function install_debian_root_docker(){
+    echo "[...]install_debian_root_docker......."
+    
+    # 检查是否debian
+    if [ ! -f /etc/debian_version ]; then
+        echo "not debian"
+        return
+    fi
+    
     sudo apt update
     sudo apt install --no-install-recommends -y gpg curl
     curl -sS https://download.docker.com/linux/debian/gpg --connect-timeout 2 --max-time 2 | gpg --dearmor > /usr/share/keyrings/docker-ce.gpg
@@ -19,7 +26,7 @@ function install_root_docker(){
         return
     fi
     sudo apt install --no-install-recommends -y docker-ce docker-ce-cli containerd.io docker-compose-plugin
-    echo "[SUC]install_root_docker"
+    echo "[SUC]install_debian_root_docker"
 }
 
-install_root_docker
+install_debian_root_docker
